requirements.txt
tk
datetime
ollama
python-docx
docx2pdf
